# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kitelepítési Támogató Rendszer API",
        "description": "QR-kódos kitelepítési, regisztrációs és befogadóhelyi nyomon követő prototípus REST API-ja.\n\nAz autentikáció Laravel Sanctum SPA session-cookie alapú: előbb GET /sanctum/csrf-cookie, utána POST /api/login. A böngészőn kívüli kliensek (pl. ez a Swagger UI) ezért a `/api/login` végpontot csak akkor tudják sikeresen meghívni, ha a kliens ugyanabból az origin-ből fut, ahonnan a cookie-kat is fogadja.",
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Helyi fejlesztői szerver"
        }
    ],
    "paths": {
...
        "/api/events/{event}/incidents": {
            "get": {
                "tags": [
                    "Incidents"
                ],
                "summary": "Eseményhez tartozó panaszok/rendkívüli események listája",
                "operationId": "a129fdc031d524f22f8be8600cc73ea1",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "open",
                                "resolved"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Incidensek időrendben"
                    }
                },
                "security": [
                    {
                        "sanctumSession": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Incidents"
                ],
                "summary": "Panasz/rendkívüli esemény rögzítése",
                "description": "Admin, vezető, regisztrátor vagy befogadóhelyi kezelő jogosult.",
                "operationId": "e85c6c10bc691afafb3d9994a1125a29",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category",
                                    "severity",
                                    "description"
                                ],
                                "properties": {
                                    "category": {
                                        "type": "string",
                                        "enum": [
                                            "complaint",
                                            "conflict",
                                            "security",
                                            "damage",
                                            "other"
                                        ]
                                    },
                                    "severity": {
                                        "type": "string",
                                        "enum": [
                                            "low",
                                            "medium",
                                            "high"
                                        ]
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "shelter_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "nullable": true
                                    },
                                    "person_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Létrehozott incidens"
                    },
                    "403": {
                        "description": "Nincs jogosultság"
                    }
                },
                "security": [
                    {
                        "sanctumSession": []
                    }
                ]
            }
        },
        "/api/incidents/{incident}/resolve": {
            "post": {
                "tags": [
                    "Incidents"
                ],
                "summary": "Panasz/rendkívüli esemény lezárása",
                "operationId": "1f70ebf03091ff2b98112b1437cf7ce1",
                "parameters": [
                    {
                        "name": "incident",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lezárt incidens"
                    },
                    "403": {
                        "description": "Nincs jogosultság"
                    }
                },
                "security": [
                    {
                        "sanctumSession": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctumSession": {
                "type": "apiKey",
                "description": "Laravel Sanctum SPA session-cookie autentikáció. Bejelentkezés: POST /api/login.",
                "name": "laravel_session",
                "in": "cookie"
            }
        }
    },
...
}